requirements.txt
enum34==1.1.6
future==0.16.0
gapic-google-logging-v2==0.10.1
gapic-google-pubsub-v1==0.10.1
google-cloud==0.21.1
google-cloud-bigquery==0.21.0
google-cloud-bigtable==0.21.0
google-cloud-core==0.21.0
google-cloud-datastore==0.21.0
google-cloud-dns==0.21.0
google-cloud-error-reporting==0.21.0
google-cloud-happybase==0.21.0
google-cloud-language==0.21.0
google-cloud-logging==0.21.0
google-cloud-monitoring==0.21.0
google-cloud-pubsub==0.21.0
google-cloud-resource-manager==0.21.0
google-cloud-runtimeconfig==0.21.0
google-cloud-storage==0.21.0
google-cloud-translate==0.21.0
google-cloud-vision==0.21.0
google-gax==0.14.1
googleapis-common-protos==1.5.0
grpc-google-iam-v1==0.10.1
grpc-google-logging-v2==0.10.1
grpc-google-pubsub-v1==0.10.1
grpcio==1.0.1
httplib2==0.9.2
isort==4.2.5
itsdangerous==0.24
Jinja2==2.8
lazy-object-proxy==1.2.2
MarkupSafe==0.23
mccabe==0.5.2
oauth2client==3.0.0
pep8==1.7.0
ply==3.8
prettyprint==0.1.5
protobuf==3.1.0.post1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pylint==1.6.4
requests==2.12.3
rsa==3.4.2
six==1.10.0
twitter==1.17.1
Werkzeug==0.11.11
wrapt==1.10.8
